Add Parser classes to improve use of SLAP
