+++ src/main/java/bestie/Bestie.java
@@ -1,8 +1,8 @@
 package bestie;
 
 
+import bestie.Parser.Parser;
 import bestie.command.Command;
-import bestie.command.ExitCommand;
 
 
 /**
